Add yaml safe dumper class customizing str dumping
