+++ src/middlewared/middlewared/plugins/chart_releases_linux/utils.py
@@ -1,7 +1,6 @@
 import copy
 import enum
 import os
-import yaml
 
 from middlewared.utils import run as _run
 
@@ -25,16 +24,6 @@ class Resources(enum.Enum):
     PERSISTENT_VOLUME_CLAIM = 'persistent_volume_claims'
     POD = 'pods'
     STATEFULSET = 'statefulsets'
-
-
-# We would like to customize safe dumper here so that when it dumps values, we quote strings
-# why we want to do this is for instances when strings like 'y' are treated as boolean true
-# by yaml and if we don't dump this enclosed with quotes, helm treats 'y' as true and we get inconsistent
-# usage
-def initialize_yaml_str_representer():
-    yaml.add_representer(
-        str, lambda dumper, data: dumper.represent_scalar('tag:yaml.org,2002:str', data, style='"'), yaml.SafeDumper
-    )
 
 
 def get_action_context(release_name):
